Input form validator helper
